+++ readme.md
@@ -1,0 +1,1 @@
+collision free
